Create README.md
This is a student project.
Nashville Software School is a 6 month software developer bootcamp for people with no previous experience coding.
The first three months focuses on the front-end and months 4 - 6 focus on back-end development.
During the final two weeks of the front-end part of the course each student must design and build their own web app to 
showcase some of the skills they have learned.
This is my web app; it's called MusicLync and is designed to bring musicians together. The idea is that the more people
practice with different musicians, the better they become.
The minimum viable product is supposed to allow musicians to create their own profile and to find other musicians based on searchable attributes, such as what instruments are played, owned, the musician's location, age group, influences, etc.
Musicians may post links to their own material and social media and may write reviews. They can add whether they teach and what their fees are for teaching, gigs or hiring their equipment.

The wider scope of the app would add in Bands, Venues, Managers and Industry. Each will be able to log in and create their own profiles and search for any of the other categories.

This project uses Redis, jade, less,  javascript, jQuery, express, foundation, lodash, grunt, ajax, mongoose, mongoDB, node and hosting with heroku.